+++ GastroFaza/Controllers/OrderController.cs
@@ -304,119 +304,159 @@ namespace GastroFaza.Controllers
         }
 
 
-
-
-
         [Route("PayForOrder")]
         public async Task<IActionResult> PayForOrder()
         {
-            int id = int.Parse(HttpContext.Session.GetString("current order"));
-            Order order =  await this.dbContext.Orders.FirstOrDefaultAsync(o => o.Id == id);
-            if (order == null)
+            if (HttpContext.Session.GetString("email") != null)
             {
-                throw new Exception("Order not found");
+                if (HttpContext.Session.GetString("isWorker") == "false")
+                {
+                    int id = int.Parse(HttpContext.Session.GetString("current order"));
+                    Order order =  await this.dbContext.Orders.FirstOrDefaultAsync(o => o.Id == id);
+                    if (order == null)
+                    {
+                        throw new Exception("Order not found");
+                    }
+                    order.Status = Status.Przyjete;
+
+                    string userEmail = HttpContext.Session.GetString("email");
+                    var client = await this.dbContext.Clients.FirstOrDefaultAsync(u => u.Email == userEmail);
+
+                    var dishOrder = this.dbContext.DishOrders.Where(x => x.OrderId == id);
+                    var dishList = await this.dbContext.Dishs.ToListAsync();
+
+                    History historyOrder = new History();
+                    historyOrder.Date = DateTime.Now;
+                    historyOrder.Stars = 0;
+
+                    foreach (var x in dishOrder)
+                    {
+                        var dish = dishList.FirstOrDefault(d => d.Id == x.DishesId);
+                        historyOrder.Dishes += dish.Name.ToString() + ", ";
+                    }
+
+                    historyOrder.AddedById = client.Id;
+
+                    this.dbContext.Histories.Add(historyOrder);
+
+                    try
+                    {
+                        await this.dbContext.SaveChangesAsync();
+                    }
+                    catch (DbUpdateException e)
+                    {
+                        throw new DbUpdateException("Error DataBase", e);
+                    }
+                    HttpContext.Session.SetString("current order", String.Empty);
+                    return View();
+                }
+                return Forbid();
             }
-            order.Status = Status.Przyjete;
-
-            string userEmail = HttpContext.Session.GetString("email");
-            var client = await this.dbContext.Clients.FirstOrDefaultAsync(u => u.Email == userEmail);
-
-            var dishOrder = this.dbContext.DishOrders.Where(x => x.OrderId == id);
-            var dishList = await this.dbContext.Dishs.ToListAsync();
-
-            History historyOrder = new History();
-            historyOrder.Date = DateTime.Now;
-            historyOrder.Stars = 0;
-
-            foreach (var x in dishOrder)
-            {
-                var dish = dishList.FirstOrDefault(d => d.Id == x.DishesId);
-                historyOrder.Dishes += dish.Name.ToString() + ", ";
-            }
-
-            historyOrder.AddedById = client.Id;
-
-            this.dbContext.Histories.Add(historyOrder);
-
-            try
-            {
-                await this.dbContext.SaveChangesAsync();
-            }
-            catch (DbUpdateException e)
-            {
-                throw new DbUpdateException("Error DataBase", e);
-            }
-            HttpContext.Session.SetString("current order", String.Empty);
-            return View();
-        }
+            return RedirectToAction("Login", "Account");
+}
 
 
         public async Task<IActionResult> RemoveDishFromOrder(Dish dish)
         {
-
-            int id = int.Parse(HttpContext.Session.GetString("current order"));
-
-            var dishOrder = await this.dbContext.DishOrders.SingleOrDefaultAsync(x => x.OrderId == id && x.DishesId == dish.Id);
-
-            this.dbContext.DishOrders.Remove(dishOrder);
-
-            try
+            if (HttpContext.Session.GetString("email") != null)
             {
-                await this.dbContext.SaveChangesAsync();
-            }
-            catch (DbUpdateException e)
-            {
-                throw new DbUpdateException("Error DataBase", e);
-            }
+                if (HttpContext.Session.GetString("Role") != "Cook")
+                {
 
-            return RedirectToAction("Order");
+                    int id = int.Parse(HttpContext.Session.GetString("current order"));
+
+                    var dishOrder = await this.dbContext.DishOrders.SingleOrDefaultAsync(x => x.OrderId == id && x.DishesId == dish.Id);
+
+                    this.dbContext.DishOrders.Remove(dishOrder);
+
+                    try
+                    {
+                        await this.dbContext.SaveChangesAsync();
+                    }
+                    catch (DbUpdateException e)
+                    {
+                        throw new DbUpdateException("Error DataBase", e);
+                    }
+
+                    return RedirectToAction("Order");
+                }
+                return Forbid();
+            }
+            return RedirectToAction("Login", "Account");
         }
 
         [Route("Create")]
         public async Task<IActionResult> Create()
         {
-            var order = new Order();
+            if (HttpContext.Session.GetString("email") != null)
+            {
+                if (HttpContext.Session.GetString("Role") != "Cook")
+                {
+                    var order = new Order();
             
-            if (await this.dbContext.Workers.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email")) != null)
-            {
-                var user = await this.dbContext.Workers.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email"));
-                order.AddedById = user.Id;
-            }
-            else if(await this.dbContext.Clients.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email")) != null)
-            {
-                var user = await this.dbContext.Clients.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email"));
-                order.AddedById = user.Id;
-            } else
-            {
-                return RedirectToAction("Login", "Account");
-            }
+                    if (await this.dbContext.Workers.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email")) != null)
+                    {
+                        var user = await this.dbContext.Workers.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email"));
+                        order.AddedById = user.Id;
+                    }
+                    else if(await this.dbContext.Clients.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email")) != null)
+                    {
+                        var user = await this.dbContext.Clients.FirstOrDefaultAsync(u => u.Email == HttpContext.Session.GetString("email"));
+                        order.AddedById = user.Id;
+                    } else
+                    {
+                        return RedirectToAction("Login", "Account");
+                    }
 
-            order.Status = Status.Nowe;
+                    order.Status = Status.Nowe;
 
-            this.dbContext.Orders.Add(order);
-            try
-            {
-                await this.dbContext.SaveChangesAsync();
-                HttpContext.Session.SetString("current order",this.dbContext.Orders.FirstOrDefault(u => u == order).Id.ToString());
+                    this.dbContext.Orders.Add(order);
+                    try
+                    {
+                        await this.dbContext.SaveChangesAsync();
+                        HttpContext.Session.SetString("current order",this.dbContext.Orders.FirstOrDefault(u => u == order).Id.ToString());
+                    }
+                    catch (DbUpdateException e)
+                    {
+                        throw new DbUpdateException("Error DataBase", e);
+                    }
+                    return RedirectToAction("GetAll", "Dish");
+                }
+                return Forbid();
             }
-            catch (DbUpdateException e)
-            {
-                throw new DbUpdateException("Error DataBase", e);
-            }
-            return RedirectToAction("GetAll", "Dish");
-        }
+            return RedirectToAction("Login", "Account");
+}
         public async Task<IActionResult> GetAllOrders()
         {
-            var orders = await  this.dbContext.Orders.ToListAsync();
+            if (HttpContext.Session.GetString("email") != null)
+            {
+                if (HttpContext.Session.GetString("isWorker") == "true")
+                {
+                    var orders = await  this.dbContext.Orders.ToListAsync();
 
-            return View(orders);
+                    return View(orders);
+                }
+                return Forbid();
+            }
+            return RedirectToAction("Login", "Account");
         }
+
         public async Task<IActionResult> Edit(int Id)
         {
-            var order = await this.dbContext.Orders.FirstOrDefaultAsync(s => s.Id == Id);
+            if (HttpContext.Session.GetString("email") != null)
+            {
+                if (HttpContext.Session.GetString("isWorker") == "true" && HttpContext.Session.GetString("Role") != "Cook")
+                {
+                    var order = await this.dbContext.Orders.FirstOrDefaultAsync(s => s.Id == Id);
 
-            return View(order);
+                    return View(order);
+                }
+                return Forbid();
+            }
+            return RedirectToAction("Login", "Account");
         }
+
+
         [HttpPost]
         public async Task<IActionResult> Edit(int? id, OrderDto modelDTO)
         {
@@ -442,24 +482,33 @@ namespace GastroFaza.Controllers
 
             return View(modelDTO);
         }
+
         public async Task<IActionResult> Delete(int? id)
         {
-            var order = await this.dbContext.Orders.FindAsync(id);
-            if (order == null)
+            if (HttpContext.Session.GetString("email") != null)
             {
-                return NotFound();
-            }
+                if (HttpContext.Session.GetString("isWorker") == "true" && HttpContext.Session.GetString("Role") != "Cook")
+                {
+                    var order = await this.dbContext.Orders.FindAsync(id);
+                    if (order == null)
+                    {
+                        return NotFound();
+                    }
 
-            this.dbContext.Orders.Remove(order);
-            try
-            {
-                await this.dbContext.SaveChangesAsync();
+                    this.dbContext.Orders.Remove(order);
+                    try
+                    {
+                        await this.dbContext.SaveChangesAsync();
+                    }
+                    catch (DbUpdateException e)
+                    {
+                        throw new DbUpdateException("Error DataBase", e);
+                    }
+                    return RedirectToAction("GetAllOrders");
+                }
+                return Forbid();
             }
-            catch (DbUpdateException e)
-            {
-                throw new DbUpdateException("Error DataBase", e);
-            }
-            return RedirectToAction("GetAllOrders");
+            return RedirectToAction("Login", "Account");
         }
 
 
@@ -496,7 +545,6 @@ namespace GastroFaza.Controllers
                 string userEmail = HttpContext.Session.GetString("email");
                 var client =  await this.dbContext.Clients.FirstOrDefaultAsync(u => u.Email == userEmail);
 
-
                 var clientHist = this.dbContext.Histories.Where(x => x.AddedById == client.Id);
 
                 foreach (var del in clientHist)
@@ -525,7 +573,7 @@ namespace GastroFaza.Controllers
         {
             if (HttpContext.Session.GetString("email") != null)
             {
-                if(HttpContext.Session.GetString("isWorker") != "true")
+                if (HttpContext.Session.GetString("isWorker") != "true")
                 {
                     return View();
                 }
